pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mrconfig"
dynamic = ["version"]
description = "多路径配置文件加载器"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
  "pyyaml==6.0.2",
]

[tool.hatch.version]
source = "code"
path = "mrconfig/__version__.py"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["mrconfig"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
show_error_codes = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
strict = false
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def"]
